+++ oa_tcp_manager/src/com/china/center/oa/tcp/bean/TcpIbReportItemBean.java
@@ -55,12 +55,12 @@ public class TcpIbReportItemBean implements Serializable
     /**
      * 中收金额
      */
-    private long ibMoney = 0;
+    private double ibMoney = 0.0d;
 
     /**
      * 激励金额
      */
-    private long motivationMoney = 0;
+    private double motivationMoney = 0.0d;
 
 
 
@@ -139,19 +139,19 @@ public class TcpIbReportItemBean implements Serializable
         this.amount = amount;
     }
 
-    public long getIbMoney() {
+    public double getIbMoney() {
         return ibMoney;
     }
 
-    public void setIbMoney(long ibMoney) {
+    public void setIbMoney(double ibMoney) {
         this.ibMoney = ibMoney;
     }
 
-    public long getMotivationMoney() {
+    public double getMotivationMoney() {
         return motivationMoney;
     }
 
-    public void setMotivationMoney(long motivationMoney) {
+    public void setMotivationMoney(double motivationMoney) {
         this.motivationMoney = motivationMoney;
     }
